+++ packages/firestore/src/core/target.ts
@@ -40,6 +40,7 @@ import {
 } from '../model/values';
 import { Value as ProtoValue } from '../protos/firestore_proto_api';
 import { debugAssert, debugCast, fail } from '../util/assert';
+import { SortedSet } from '../util/sorted_set';
 import { isNullOrUndefined } from '../util/types';
 
 /**
@@ -306,12 +307,8 @@ export function targetGetLowerBound(
   for (const segment of fieldIndexGetDirectionalSegments(fieldIndex)) {
     const segmentBound =
       segment.kind === IndexKind.ASCENDING
-        ? targetGetLowerBoundForField(target, segment.fieldPath, target.startAt)
-        : targetGetUpperBoundForField(
-            target,
-            segment.fieldPath,
-            target.startAt
-          );
+        ? targetGetAscendingBound(target, segment.fieldPath, target.startAt)
+        : targetGetDescendingBound(target, segment.fieldPath, target.startAt);
 
     values.push(segmentBound.value);
     inclusive &&= segmentBound.inclusive;
@@ -336,8 +333,8 @@ export function targetGetUpperBound(
   for (const segment of fieldIndexGetDirectionalSegments(fieldIndex)) {
     const segmentBound =
       segment.kind === IndexKind.ASCENDING
-        ? targetGetUpperBoundForField(target, segment.fieldPath, target.endAt)
-        : targetGetLowerBoundForField(target, segment.fieldPath, target.endAt);
+        ? targetGetDescendingBound(target, segment.fieldPath, target.endAt)
+        : targetGetAscendingBound(target, segment.fieldPath, target.endAt);
 
     values.push(segmentBound.value);
     inclusive &&= segmentBound.inclusive;
@@ -350,7 +347,7 @@ export function targetGetUpperBound(
  * Returns the value to use as the lower bound for ascending index segment at
  * the provided `fieldPath` (or the upper bound for an descending segment).
  */
-function targetGetLowerBoundForField(
+function targetGetAscendingBound(
   target: Target,
   fieldPath: FieldPath,
   bound: Bound | null
@@ -423,9 +420,9 @@ function targetGetLowerBoundForField(
 
 /**
  * Returns the value to use as the upper bound for ascending index segment at
- * the provided `fieldPath` (or the lower bound for an descending segment).
+ * the provided `fieldPath` (or the lower bound for a descending segment).
  */
-function targetGetUpperBoundForField(
+function targetGetDescendingBound(
   target: Target,
   fieldPath: FieldPath,
   bound: Bound | null
@@ -494,6 +491,46 @@ function targetGetUpperBoundForField(
   }
 
   return { value, inclusive };
+}
+
+/** Returns the number of segments of a perfect index for this target. */
+export function targetGetSegmentCount(target: Target): number {
+  let fields = new SortedSet<FieldPath>(FieldPath.comparator);
+  let hasArraySegment = false;
+
+  for (const filter of target.filters) {
+    // TODO(orquery): Use the flattened filters here
+    const fieldFilter = filter as FieldFilter;
+
+    // __name__ is not an explicit segment of any index, so we don't need to
+    // count it.
+    if (fieldFilter.field.isKeyField()) {
+      continue;
+    }
+
+    // ARRAY_CONTAINS or ARRAY_CONTAINS_ANY filters must be counted separately.
+    // For instance, it is possible to have an index for "a ARRAY a ASC". Even
+    // though these are on the same field, they should be counted as two
+    // separate segments in an index.
+    if (
+      fieldFilter.op === Operator.ARRAY_CONTAINS ||
+      fieldFilter.op === Operator.ARRAY_CONTAINS_ANY
+    ) {
+      hasArraySegment = true;
+    } else {
+      fields = fields.add(fieldFilter.field);
+    }
+  }
+
+  for (const orderBy of target.orderBy) {
+    // __name__ is not an explicit segment of any index, so we don't need to
+    // count it.
+    if (!orderBy.field.isKeyField()) {
+      fields = fields.add(orderBy.field);
+    }
+  }
+
+  return fields.size + (hasArraySegment ? 1 : 0);
 }
 
 export abstract class Filter {